split emGUI glue and UI
