+++ emGUI-example/GUI.h
@@ -2,15 +2,8 @@
 #define __GUI_H
 
 #include "emGUI/emGUI.h"
+
 #include <windows.h>
-#include <objidl.h>
-#include <gdiplus.h>
-#include <iostream>
-
-using namespace Gdiplus;
-
-#define SCREEN_WIDTH EMGUI_LCD_WIDTH
-#define SCREEN_HEIGHT EMGUI_LCD_HEIGHT
 
 typedef struct {
 	float averageCurrent;
@@ -27,17 +20,13 @@ typedef enum {
 
 //events
 
-ARGB convertColor(uint16_t color);
 bool bGUI_InitInterfce();
 bool bGUIOnWindowCloseHandlerMain(xWidget *);
 bool bGUIOnWindowCloseHandler(xWidget *);
-void vGUIsetCurrentHDC(Graphics * gr);
-void vGUIpaintEventHandler();
 void vGUIpushClickHandler(LPARAM lParam);
 void vGUIpopClickHandler(LPARAM lParam);
-void vGUIeraseBackgroudHandler();
 void vGUIUpdateCurrentMonitor();
-void vGUIWriteToLabel(char * msg);
+bool bGUIonInterfaceCreateHandler(xWidget *);
 
 xPlotData_t * pxGUIGetPlotData();
 extraParams_t * pxGUIGetExtraParams();
